// ssr/enums/mixpanel-people.enum.ts
export enum MixpanelPeopleAction {
    SET = 'set',
    APPEND = 'append',
    INCREMENT = 'increment',
    SET_ONCE = 'set_once',
    REMOVE = 'remove',
    UNSET = 'unset',
    UNION = 'union',
    CLEAR_CHANGES = 'clear_charges',
    DELETE_USER = 'delete_user',
    TRACK_CHARGE = 'track_charge'
}